fix #17402
