--- .ai-sdk-factory/reproduction-replay.json
@@ -1,12 +1,0 @@
-{
-  "version": 1,
-  "issueId": 17402,
-  "command": "pnpm -C examples/ai-functions exec tsx src/reproduction/issue-17402-tool-search-name-collision.ts",
-  "artifactPaths": [
-    "examples/ai-functions/src/reproduction/issue-17402-tool-search-name-collision.ts"
-  ],
-  "expectedFailure": {
-    "exitCode": 1,
-    "signal": "ISSUE_17402_REPRODUCED: regular function tool_search was replayed as tool_search_call without arguments and OpenAI rejected the request"
-  }
-}
