+++ .changeset/calm-tools-replay.md
@@ -1,0 +1,5 @@
+---
+'@ai-sdk/openai': patch
+---
+
+Replay regular functions named `tool_search` as function calls in Responses API history.
