--- packages/openai/src/responses/__fixtures__/openai-issue-17402-error.1.json
@@ -1,8 +1,0 @@
-{
-  "error": {
-    "message": "Missing required parameter: 'input[1].arguments'.",
-    "type": "invalid_request_error",
-    "param": "input[1].arguments",
-    "code": "missing_required_parameter"
-  }
-}
